requirements.txt
# MCP Server
mcp>=1.0.0

# LLM (Gemini via OpenAI-kompatibelt API)
openai>=1.0.0

# Google APIs
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Utilities
python-dotenv>=1.0.0
